Cleaned up code for API routes
